refactor: 責務分離とSRPの適用 (Phase 4)
- useGameState.ts を4つのフック + Facadeに分割
  - useGameSetup: マップ生成・プレイヤー初期化・敵スポーン・ギミック配置
  - useScreenTransition: 画面遷移ハンドラー
  - useStageManagement: ステージ進行・報酬・引き継ぎ
  - useGameAudio: BGM/SE管理・画面ごとの切り替え
- Game.tsx を5つのコンポーネントに分割
  - GameModals: ClassSelectScreen, LevelUpOverlay, HelpOverlay
  - GameHUD: HP/レベル/ステータス/タイマー等の情報表示
  - GameControls: D-pad・キーボード入力・連続移動
  - GameCanvas: Canvas要素ラッパー (React.memo)
- tickGameState.ts から3つのユースケースを抽出
  - resolveTraps: 罠トリガー処理
  - resolveRegen: リジェネ処理
  - resolveEnemyUpdates: 敵更新・死亡フィルタ
- useGameLoop.ts から useEffectDispatcher.ts を抽出
- resolvePlayerDamage に actualDamage を伝播 (Phase 3レビュー残課題)
- MATH_RANDOM_PROVIDER / SYSTEM_CLOCK_PROVIDER シングルトンを削除
- tickGameState に RandomProvider を伝播（テレポート罠対応）
