+++ src/features/ipne/infrastructure/index.ts
@@ -1,9 +1,9 @@
 export { BROWSER_ENV_PROVIDER } from './browser/BrowserEnvProvider';
 export type { BrowserEnvProvider } from './browser/BrowserEnvProvider';
 
-export { SYSTEM_CLOCK_PROVIDER, DateClockProvider } from './clock/ClockProvider';
+export { DateClockProvider } from './clock/ClockProvider';
 
-export { MATH_RANDOM_PROVIDER, MathRandomProvider } from './random/RandomProvider';
+export { MathRandomProvider } from './random/RandomProvider';
 
 export { SequentialIdGenerator } from './id/SequentialIdGenerator';
