+++ src/features/ipne/presentation/index.ts
@@ -7,9 +7,14 @@ export { CONFIG, PROLOGUE_TEXTS } from './config';
 export { TitleScreen, AudioSettingsComponent } from './screens/Title';
 export { PrologueScreen } from './screens/Prologue';
 export { GameScreen, ClassSelectScreen, LevelUpOverlayComponent, HelpOverlayComponent } from './screens/Game';
+export type { EffectEvent } from './screens/Game';
+export { GameHUD } from './screens/GameHUD';
+export { GameControls } from './screens/GameControls';
+export { GameCanvas } from './screens/GameCanvas';
 export { ClearScreen, GameOverScreen } from './screens/Clear';
 
 // カスタムフック
 export { useGameState } from './hooks/useGameState';
 export type { GameState } from './hooks/useGameState';
 export { useGameLoop } from './hooks/useGameLoop';
+export { useEffectDispatcher } from './hooks/useEffectDispatcher';
